Miscellaneous perf improvements * Only track cache metrics in debug builds * Simplified MoveSet, added EmptySet * Reduced size of Position by defaulting to Q/R

+++ Mzinga.Shared/Core/ZobristHash.cs
@@ -4,7 +4,7 @@
 // Author:
 //       Jon Thysell <thysell@gmail.com>
 // 
-// Copyright (c) 2018 Jon Thysell <http://jonthysell.com>
+// Copyright (c) 2018, 2019 Jon Thysell <http://jonthysell.com>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -91,6 +91,6 @@ namespace Mzinga.Core
 
         public const long EmptyBoard = 0;
 
-        private const int NumUniquePositions = Position.MaxStack * EnumUtils.NumPieceNames * EnumUtils.NumPieceNames;
+        private const int NumUniquePositions = (int)Position.MaxStack * EnumUtils.NumPieceNames * EnumUtils.NumPieceNames;
     }
 }
